refactor(transport): expose Result types on public factory.transport surface
Move Err/Ok/Result/SanitizedError/InboxStream to factory.transport public API
and update codecs, outbound helpers, and adapters to import from there.
Remove obsolete importlinter ignore_imports; update AGENTS.md and CURRENT.generated.md.
