+++ src/factory/llm/cli_nats_codec.py
@@ -19,7 +19,7 @@ from pydantic import ValidationError
 from factory.core.envelope_fields import mint_work_envelope_fields
 from factory.core.messaging.events import LlmEvent, ResultLlmEvent, TextLlmEvent
 from factory.core.ports.llm import LlmResult
-from factory.transport._result import Err, Result, SanitizedError
+from factory.transport import Err, Result, SanitizedError
 from roxabi_contracts.errors import KNOWN_CODES, WorkerError
 from roxabi_contracts.llm import LlmChunkEvent, LlmRequest, LlmResponse
