+++ src/factory/nats/audio/nats_tts_codec.py
@@ -16,7 +16,7 @@ from pydantic import ValidationError
 
 from factory.core.ports.tts import SynthesisResult
 from factory.nats.envelope_fields import mint_work_envelope_fields
-from factory.transport._result import Err, Result, SanitizedError
+from factory.transport import Err, Result, SanitizedError
 from roxabi_contracts.voice import TtsRequest, TtsResponse
 from roxabi_contracts.voice.constants import TTS_CONFIG_FIELDS
