+++ src/factory/transport/__init__.py
@@ -1,5 +1,10 @@
-"""NATS transport layer (Phase 1 #1278)."""
+"""NATS transport layer (Phase 1 #1278).
 
-from ._result import SanitizedError
+Public surface for the Result helper (Err/Ok/Result/SanitizedError/InboxStream).
+Prefer importing from here, never factory.transport._result directly (see
+.importlinter transport-result-public-surface contract).
+"""
 
-__all__ = ["SanitizedError"]
+from ._result import Err, InboxStream, Ok, Result, SanitizedError
+
+__all__ = ["Err", "InboxStream", "Ok", "Result", "SanitizedError"]
